Convert transaction committer to applicative style
