+++ ledger/participant-state/kvutils/src/main/scala/com/daml/ledger/participant/state/kvutils/committer/Committer.scala
@@ -5,14 +5,15 @@ package com.daml.ledger.participant.state.kvutils.committer
 
 import com.codahale.metrics.Timer
 import com.daml.ledger.participant.state.kvutils.DamlKvutils.{
+  DamlConfigurationEntry,
   DamlLogEntry,
   DamlLogEntryId,
   DamlStateKey,
   DamlStateValue
 }
-import com.daml.ledger.participant.state.kvutils.DamlStateMap
+import com.daml.ledger.participant.state.kvutils.{Conversions, DamlStateMap, Err}
 import com.daml.ledger.participant.state.kvutils.committer.Committer._
-import com.daml.ledger.participant.state.v1.ParticipantId
+import com.daml.ledger.participant.state.v1.{Configuration, ParticipantId}
 import com.daml.lf.data.Time
 import com.daml.metrics.Metrics
 import org.slf4j.{Logger, LoggerFactory}
@@ -93,4 +94,27 @@ private[committer] trait Committer[Submission, PartialResult] {
 
 object Committer {
   type StepInfo = String
+
+  def getCurrentConfiguration(
+      defaultConfig: Configuration,
+      inputState: Map[DamlStateKey, Option[DamlStateValue]],
+      logger: Logger): (Option[DamlConfigurationEntry], Configuration) =
+    inputState
+      .getOrElse(
+        Conversions.configurationStateKey,
+        /* If we're retrieving configuration, we require it to at least
+         * have been declared as an input by the submitter as it is used
+         * to authorize configuration changes. */
+        throw Err.MissingInputState(Conversions.configurationStateKey)
+      )
+      .flatMap { v =>
+        val entry = v.getConfigurationEntry
+        Configuration
+          .decode(entry.getConfiguration)
+          .fold({ err =>
+            logger.error(s"Failed to parse configuration: $err, using default configuration.")
+            None
+          }, conf => Some(Some(entry) -> conf))
+      }
+      .getOrElse(None -> defaultConfig)
 }
